Hoist contract components into --lib
