+++ packages/--lib/abis/index.ts
@@ -1,17 +1,17 @@
 import SingleTokenRewardDistributor from './SingleTokenRewardDistributor'
 import YearnBoostedStaker from './YearnBoostedStaker'
 import yPrisma from './yPrisma'
-import Vault from './Vault'
+import VaultV2 from './VaultV2'
+import VaultV3 from './VaultV3'
 import Utilities from './Utilities'
-import OldStaker from './OldStaker'
 
 const abis = {
   SingleTokenRewardDistributor,
   YearnBoostedStaker,
   yPrisma,
-  Vault,
-  Utilities,
-  OldStaker,
+  VaultV2,
+  VaultV3,
+  Utilities
 }
 
 export default abis
