+++ packages/--lib/components/WithdrawV2.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import useData from '--lib/hooks/useData'
+import useData from '../hooks/useData'
 import InputExecute from './InputExecute'
-import env from '--lib/tools/env'
-import abis from '../app/abis'
+import env from '../tools/env'
+import abis from '../abis'
 
-export default function Withdraw({ className }: { className?: string }) {
+export default function WithdrawV2({ className }: { className?: string }) {
   const { data } = useData()
   return <InputExecute className={className}
     task={{
@@ -14,7 +14,7 @@ export default function Withdraw({ className }: { className?: string }) {
       needsApproval: false,
       parameters: {
         address: env.YPRISMA_STRATEGY,
-        abi: abis.Strategy,
+        abi: abis.VaultV2,
         functionName: 'withdraw',
         args: (amount: bigint) => [amount]
       }
