+++ packages/--lib/components/WithdrawV3.tsx
@@ -1,11 +1,11 @@
 'use client'
 
-import useData from '--lib/hooks/useData'
+import useData from '../hooks/useData'
 import InputExecute from './InputExecute'
-import env from '--lib/tools/env'
-import abis from '../app/abis'
+import env from '../tools/env'
+import abis from '../abis'
 
-export default function Withdraw({ className }: { className?: string }) {
+export default function WithdrawV3({ className }: { className?: string }) {
   const { data } = useData()
   return <InputExecute className={className}
     task={{
@@ -14,7 +14,7 @@ export default function Withdraw({ className }: { className?: string }) {
       needsApproval: false,
       parameters: {
         address: env.YPRISMA_STRATEGY,
-        abi: abis.Strategy,
+        abi: abis.VaultV3,
         functionName: 'redeem',
         args: (amount: bigint) => [amount, data.account, data.account]
       }
