--- packages/crv/app/abis/index.ts
@@ -1,17 +1,0 @@
-import SingleTokenRewardDistributor from './SingleTokenRewardDistributor'
-import YearnBoostedStaker from './YearnBoostedStaker'
-import yPrisma from './yPrisma'
-import Strategy from './Strategy'
-import Utilities from './Utilities'
-import OldStaker from './OldStaker'
-
-const abis = {
-  SingleTokenRewardDistributor,
-  YearnBoostedStaker,
-  yPrisma,
-  Strategy,
-  Utilities,
-  OldStaker,
-}
-
-export default abis
