+++ packages/crv/app/app/[tab]/page.tsx
@@ -2,7 +2,7 @@
 
 import Image from 'next/image'
 import Link from 'next/link'
-import InputBox from '../../../components/InputBox'
+import InputBox from '--lib/components/InputBox'
 import Header, { headerItems } from '../../../components/Header'
 import { useConnectModal, useAccountModal } from '@rainbow-me/rainbowkit'
 import { useAccount } from 'wagmi'
@@ -12,11 +12,11 @@ import useData from '--lib/hooks/useData'
 import Tokens from '--lib/components/Tokens'
 import ExperienceToggle from '--lib/components/ExperienceToggle'
 
-import ClaimAll from '../../../components/ClaimAll'
-import Stake from '../../../components/Stake'
-import Unstake from '../../../components/Unstake'
-import Deposit from '../../../components/Deposit'
-import Withdraw from '../../../components/Withdraw'
+import ClaimAll from '--lib/components/ClaimAll'
+import Stake from '--lib/components/Stake'
+import Unstake from '--lib/components/Unstake'
+import Deposit from '--lib/components/DepositV2'
+import Withdraw from '--lib/components/WithdrawV2'
 
 import { useContractReads } from 'wagmi'
 import { PiVaultLight } from 'react-icons/pi'
@@ -25,7 +25,7 @@ import { formatUnits } from 'viem'
 import bmath from '--lib/tools/bmath'
 import env from '--lib/tools/env'
 import Background from '../../../components/Background'
-import A from '@/components/A'
+import A from '--lib/components/A'
 import ImageOrFallback from '@/components/ImageOrFallback'
 import Zap from '@/components/Zap'
 import { useVaultContext } from '--lib/context/VaultContext'
