+++ packages/crv/app/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import Button from '../components/Button'
+import Button from '--lib/components/Button'
 import Header, { headerItems } from '../components/Header'
 import Link from 'next/link'
 import Image from 'next/image'
