--- packages/prisma/app/abis/Utilities.ts
@@ -1,3 +1,0 @@
-const abi = [{'inputs': [{'components': null, 'internalType': 'contract IYearnBoostedStaker', 'name': '_ybs', 'type': 'address'}, {'components': null, 'internalType': 'contract IRewardsDistributor', 'name': '_rewardsDistributor', 'type': 'address'}], 'stateMutability': 'nonpayable', 'type': 'constructor'}, {'inputs': [], 'name': 'MAX_STAKE_GROWTH_WEEKS', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'REWARDS_DISTRIBUTOR', 'outputs': [{'components': null, 'internalType': 'contract IRewardsDistributor', 'name': '', 'type': 'address'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'TOKEN', 'outputs': [{'components': null, 'internalType': 'contract IERC20', 'name': '', 'type': 'address'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'YBS', 'outputs': [{'components': null, 'internalType': 'contract IYearnBoostedStaker', 'name': '', 'type': 'address'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'activeRewardAmount', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'address', 'name': '_account', 'type': 'address'}, {'components': null, 'internalType': 'uint256', 'name': '_week', 'type': 'uint256'}], 'name': 'adjustedAccountWeightAt', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_week', 'type': 'uint256'}], 'name': 'adjustedGlobalWeightAt', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'address', 'name': '_account', 'type': 'address'}, {'components': null, 'internalType': 'uint256', 'name': '_week', 'type': 'uint256'}], 'name': 'getAccountStakeAmountAt', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_stakeTokenPrice', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': '_rewardTokenPrice', 'type': 'uint256'}], 'name': 'getGlobalActiveApr', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'getGlobalActiveBoostMultiplier', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_stakeTokenPrice', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': '_rewardTokenPrice', 'type': 'uint256'}], 'name': 'getGlobalMinMaxActiveApr', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': 'min', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': 'max', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_stakeTokenPrice', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': '_rewardTokenPrice', 'type': 'uint256'}], 'name': 'getGlobalMinMaxProjectedApr', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': 'min', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': 'max', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_stakeTokenPrice', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': '_rewardTokenPrice', 'type': 'uint256'}], 'name': 'getGlobalProjectedApr', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'getGlobalProjectedBoostMultiplier', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_week', 'type': 'uint256'}], 'name': 'getGlobalStakeAmountAt', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'address', 'name': '_account', 'type': 'address'}, {'components': null, 'internalType': 'uint256', 'name': '_stakeTokenPrice', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': '_rewardTokenPrice', 'type': 'uint256'}], 'name': 'getUserActiveApr', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'address', 'name': '_user', 'type': 'address'}], 'name': 'getUserActiveBoostMultiplier', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'address', 'name': '_account', 'type': 'address'}, {'components': null, 'internalType': 'uint256', 'name': '_stakeTokenPrice', 'type': 'uint256'}, {'components': null, 'internalType': 'uint256', 'name': '_rewardTokenPrice', 'type': 'uint256'}], 'name': 'getUserProjectedApr', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'address', 'name': '_user', 'type': 'address'}], 'name': 'getUserProjectedBoostMultiplier', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'getWeek', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [], 'name': 'projectedRewardAmount', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}, {'inputs': [{'components': null, 'internalType': 'uint256', 'name': '_week', 'type': 'uint256'}], 'name': 'weeklyRewardAmountAt', 'outputs': [{'components': null, 'internalType': 'uint256', 'name': '', 'type': 'uint256'}], 'stateMutability': 'view', 'type': 'function'}]
-
-export default abi
