+++ packages/prisma/app/app/[tab]/page.tsx
@@ -2,7 +2,7 @@
 
 import Image from 'next/image'
 import Link from 'next/link'
-import InputBox from '../../../components/InputBox'
+import InputBox from '--lib/components/InputBox'
 import Header, { headerItems } from '../../../components/Header'
 import { Suspense } from 'react'
 import { useConnectModal, useAccountModal } from '--lib/hooks/rainbowkit'
@@ -14,12 +14,12 @@ import useData from '--lib/hooks/useData'
 import Tokens from '--lib/components/Tokens'
 import ExperienceToggle from '--lib/components/ExperienceToggle'
 
-import ClaimAll from '../../../components/ClaimAll'
-import Stake from '../../../components/Stake'
-import Unstake from '../../../components/Unstake'
-import Mint from '../../../components/Mint'
-import Deposit from '../../../components/Deposit'
-import Withdraw from '../../../components/Withdraw'
+import ClaimAll from '--lib/components/ClaimAll'
+import Stake from '--lib/components/Stake'
+import Unstake from '--lib/components/Unstake'
+import Mint from '--lib/components/Mint'
+import Deposit from '--lib/components/DepositV3'
+import Withdraw from '--lib/components/WithdrawV3'
 import WithdrawAllFromOldStaker from '../../../components/WithdrawAllFromOldStaker'
 
 import { useContractReads } from 'wagmi'
@@ -29,7 +29,7 @@ import { formatUnits } from 'viem'
 import bmath from '--lib/tools/bmath'
 import env from '--lib/tools/env'
 import Background from '../../../components/Background'
-import A from '@/components/A'
+import A from '--lib/components/A'
 import YbsDataBox from '--lib/components/YbsDataBox'
 import VaultDataBox from '--lib/components/VaultDataBox'
 import { useTab } from '--lib/hooks/useTab'
