--- packages/prisma/components/Mint.tsx
@@ -1,22 +1,0 @@
-'use client'
-
-import useData from '--lib/hooks/useData'
-import InputExecute from './InputExecute'
-import env from '--lib/tools/env'
-import abis from '../app/abis'
-
-export default function Mint({ className }: { className?: string }) {
-  const { data } = useData()
-  return <InputExecute className={className}
-    task={{
-      verb: 'mint',
-      token: data.asset,
-      needsApproval: true,
-      parameters: {
-        address: env.YPRISMA,
-        abi: abis.yPrisma,
-        functionName: 'mint',
-        args: (amount: bigint) => [amount, data.account]
-      }
-    }} />
-}
