--- packages/prisma/components/Unstake.tsx
@@ -1,22 +1,0 @@
-'use client'
-
-import useData, { TokenSchema } from '--lib/hooks/useData'
-import InputExecute from './InputExecute'
-import env from '--lib/tools/env'
-import abis from '../app/abis'
-
-export default function Unstake({ className }: { className?: string }) {
-  const { data } = useData()
-  return <InputExecute className={className}
-    task={{
-      verb: 'unstake',
-      token: TokenSchema.parse(data.staker),
-      needsApproval: false,
-      parameters: {
-        address: env.YPRISMA_BOOSTED_STAKER,
-        abi: abis.YearnBoostedStaker,
-        functionName: 'unstake',
-        args: (amount: bigint) => [amount, data.account]
-      }
-    }} />
-}
